Move Multiboot 2 string functions to separate module
