+++ src/multiboot2/info_print.c
@@ -6,10 +6,6 @@
 
 #include <stddef.h>
 
-static const char *KernAux_Multiboot2_ITag_to_str(
-    enum KernAux_Multiboot2_ITag tag_type
-);
-
 static void KernAux_Multiboot2_ITag_MemoryMap_print(
     const struct KernAux_Multiboot2_ITag_MemoryMap *tag,
     void (*printf)(const char *format, ...) __attribute__((format(printf, 1, 2)))
@@ -19,59 +15,6 @@ static void KernAux_Multiboot2_ITag_ELFSymbols_print(
     const struct KernAux_Multiboot2_ITag_ELFSymbols *tag,
     void (*printf)(const char *format, ...) __attribute__((format(printf, 1, 2)))
 );
-
-const char *KernAux_Multiboot2_ITag_to_str(
-    enum KernAux_Multiboot2_ITag tag_type
-) {
-    switch (tag_type) {
-    case KERNAUX_MULTIBOOT2_ITAG_NONE:
-        return "none";
-    case KERNAUX_MULTIBOOT2_ITAG_BOOT_CMD_LINE:
-        return "boot cmd line";
-    case KERNAUX_MULTIBOOT2_ITAG_BOOT_LOADER_NAME:
-        return "boot loader name";
-    case KERNAUX_MULTIBOOT2_ITAG_MODULE:
-        return "module";
-    case KERNAUX_MULTIBOOT2_ITAG_BASIC_MEMORY_INFO:
-        return "basic memory info";
-    case KERNAUX_MULTIBOOT2_ITAG_BIOS_BOOT_DEVICE:
-        return "BIOS boot device";
-    case KERNAUX_MULTIBOOT2_ITAG_MEMORY_MAP:
-        return "memory map";
-    case KERNAUX_MULTIBOOT2_ITAG_VBE_INFO:
-        return "VBE info";
-    case KERNAUX_MULTIBOOT2_ITAG_FRAMEBUFFER_INFO:
-        return "framebuffer info";
-    case KERNAUX_MULTIBOOT2_ITAG_ELF_SYMBOLS:
-        return "ELF symbols";
-    case KERNAUX_MULTIBOOT2_ITAG_APM_TABLE:
-        return "APM table";
-    case KERNAUX_MULTIBOOT2_ITAG_EFI_32BIT_SYSTEM_TABLE_PTR:
-        return "EFI 32bit system table ptr";
-    case KERNAUX_MULTIBOOT2_ITAG_EFI_64BIT_SYSTEM_TABLE_PTR:
-        return "EFI 64bit system table ptr";
-    case KERNAUX_MULTIBOOT2_ITAG_SMBIOS_TABLES:
-        return "SMBIOS tables";
-    case KERNAUX_MULTIBOOT2_ITAG_ACPI_OLD_RSDP:
-        return "ACPI old RSDP";
-    case KERNAUX_MULTIBOOT2_ITAG_ACPI_NEW_RSDP:
-        return "ACPI new RSDP";
-    case KERNAUX_MULTIBOOT2_ITAG_NETWORKING_INFO:
-        return "networking info";
-    case KERNAUX_MULTIBOOT2_ITAG_EFI_MEMORY_MAP:
-        return "EFI memory map";
-    case KERNAUX_MULTIBOOT2_ITAG_EFI_BOOT_SERVICES_NOT_TERMINATED:
-        return "EFI boot services not terminated";
-    case KERNAUX_MULTIBOOT2_ITAG_EFI_32BIT_IMAGE_HANDLE_PTR:
-        return "EFI 32bit image handle ptr";
-    case KERNAUX_MULTIBOOT2_ITAG_EFI_64BIT_IMAGE_HANDLE_PTR:
-        return "EFI 64bit image handle ptr";
-    case KERNAUX_MULTIBOOT2_ITAG_IMAGE_LOAD_BASE_PHYS_ADDR:
-        return "image load base phys addr";
-    default:
-        return NULL;
-    }
-}
 
 void KernAux_Multiboot2_Info_print(
     const struct KernAux_Multiboot2_Info *const multiboot2_info,
